+++ src/processes/prepare_mira_reports.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code, unused_imports)]
+use crate::utils::dataframes::*;
 use clap::Parser;
 use csv::ReaderBuilder;
 use either::Either;
@@ -109,45 +110,6 @@ fn read_yaml<R: std::io::Read>(reader: R) -> Result<QCConfig, Box<dyn std::error
     Ok(config)
 }
 
-fn coverage_df(irma_path: &PathBuf) -> Result<DataFrame, Box<dyn Error>> {
-    // Define the pattern to match text files
-    let pattern = format!(
-        "{}/*/IRMA/*/tables/*coverage.txt",
-        irma_path.to_string_lossy()
-    );
-    println!("{}", pattern);
-
-    // Initialize an empty DataFrame to hold the combined data
-    let mut combined_cov_df: Option<DataFrame> = None;
-
-    // Iterate over all files matching the pattern
-    for entry in glob(&pattern).expect("Failed to read glob pattern") {
-        match entry {
-            Ok(path) => {
-                // Read the CSV file into a DataFrame
-                let file_path = path.to_str().unwrap();
-                println!("Reading file: {}", file_path);
-
-                let df = CsvReader::from_path(file_path)?.has_header(true).finish()?;
-
-                // Combine the DataFrame with the existing one
-                combined_cov_df = match combined_cov_df {
-                    Some(existing_df) => Some(existing_df.vstack(&df)?),
-                    None => Some(df),
-                };
-            }
-            Err(e) => println!("Error reading file: {}", e),
-        }
-    }
-
-    // Return the combined DataFrame or an error if no data was found
-    if let Some(df) = combined_cov_df {
-        Ok(df)
-    } else {
-        Err("No files found or no data to combine.".into())
-    }
-}
-
 pub fn prepare_mira_reports_process(args: ReportsArgs) -> Result<(), Box<dyn Error>> {
     // Read in samplesheet
     let samplesheet = read_csv_to_dataframe(&args.samplesheet)?;
@@ -157,11 +119,18 @@ pub fn prepare_mira_reports_process(args: ReportsArgs) -> Result<(), Box<dyn Err
     let qc_config: QCConfig = read_yaml(qc_yaml_path)?;
 
     //cov df
-    let cov_df = coverage_df(&args.irma_path)?;
+    let mut cov_df = coverage_df(&args.irma_path)?;
 
-    //println!("{:?}", samplesheet);
-    println!("{:?}", qc_config);
+    /*
+    let output_file = "./output_df.csv"; // Adjust the path as needed
+
+    CsvWriter::new(std::fs::File::create(output_file)?)
+        .has_header(true)
+        .finish(&mut cov_df)?;
+     */
+
     println!("{:?}", samplesheet);
+    println!("{:?}", qc_config);
     println!("{:?}", cov_df);
 
     Ok(())
